requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonstream==0.0.1
MarkupSafe==2.1.1
psycopg2==2.9.3
PyYAML==6.0
urllib3==1.26.9
Werkzeug==2.2.2
zipp==3.8.1